[refactor] split modules, fix auth
